Fix failing `cider-expected-ns` test on windows
In cider-expected-ns, a test was providing a `path` of
"/a/foo/bar/baz_utils.clj" when `cider-classpath-entries had a value of
`'("/a" "/b" "/c" "/c/inner" "/base/clj" "/base/clj-dev")`

The tests expects the cider-classpath to be picked up in
`cider-expected-ns` and drop the `/a` prefix from the generated
namespaced, leaving `foo.bar.baz-utils`

This was not happening on windows because cider-expected-ns was first
calling `file-truename` on `path`, which prefixed the path with the
drive name. This caused the `file-in-directory-p` check to start failing
because "z:/a/foo/..." is not considered to be in `"/a"` on windows.
This made cider think that the `path` was not located anywhere on the
current classpath.

Mapping over all the directories with `file-truename` fixes this on
windows, and should also be a safe operation on sane operating systems.

I am not sure how clojure-ts-mode changes caused this bug, or surfaced
it.